Add a few methods in Calibration/IsolatedParticles
